### 2022-07-10 #549 @markean
Presubmission inquiry - melt: Multiple Empirical Likelihood Tests
Submitting Author Name: Eunseop Kim
Submitting Author Github Handle: @markean
Repository:  https://github.com/markean/melt
Submission type: Pre-submission

---

```
Package: melt
Title: Multiple Empirical Likelihood Tests
Version: 1.6.0
    person("Eunseop", "Kim", , "kim.7302@osu.edu", role = c("aut", "cre")),
    person("Steven", "MacEachern", role = c("ctb", "ths")),
    person("Mario", "Peruggia", role = c("ctb", "ths"))
  )
Description: Performs multiple empirical likelihood tests for linear and
    generalized linear models. The core computational routines are
    implemented using the 'Eigen' C++ library and 'RcppEigen' interface,
    with OpenMP for parallel computation. Details of multiple testing
    procedures are given in Kim, MacEachern, and Peruggia (2021)
    <arxiv:2112.09206>.
    graphics,
    Rcpp,
    stats
    covr,
    microbenchmark,
    withr
    BH,
    dqrng,
    Rcpp,
    RcppEigen
    knitr
NeedsCompilation: yes
Roxygen: list (markdown = TRUE, roclets = c ("namespace", "rd",
    "srr::srr_stats_roclet"))

```

The package performs hypothesis testing with empirical likelihood for linear models and generalized linear models.

Yes.

Academic statisticians who are interested in empirical likelihood-based inference for (generalized) linear models.

No, at least in my understanding.

Yes,

How long does the peer-review process take on average? The plan is to submit the package with a manuscript to the Journal of Statistical Software.
