### 2022-07-10 #549 @emilyriederer
Hi @markean ! Thank you for your submission to rOpenSci

I'm tagging in @mpadge to answer your question on the review timeline for statistical packages and interactions with JSS
