### 2022-07-12 #549 @mpadge
Thanks @emilyriederer, and note that we'll soon have a new command for stats submissions, `@ropensci-review-bot check srr`, to check compliances with standards. In the meantime, this package gives this result:

---

## 'srr' standards compliance:

- complied with: 86 / 116 = 74.1%
- not complied with: 30 / 116 = 25.9%

This package complies with > 50% of all standads and may be submitted.

---

@markean Some responses to your questions:

> How long does the peer-review process take on average?

For standard (non-stats) submissions, we generally aim to review packages within 2-3 months. The statistical software review process is still very much a work-in-progress, and requires quite a bit more effort on the part of all involved. This means reviews thus far are taking longer than conventional reviews, and the entire process is definitely taking longer than the standard 2-3 months. It is nevertheless early days, and we will endeavour to find out why and how the process is slower than for non-stats reviews, and will do our best to speed things up. I'd suggest ou should at present expect the process to take vaguely between 3 and 6 months.

> The plan is to submit the package with a manuscript to the Journal of Statistical Software.

That's great news! We have two members of the editorial borad of JSS as editors of our statistical peer review program (ping @rkillick & @tdhock :smile:). A slightly longer-term aim is for rOpenSci statistical peer-review to lead to expedited review processes in JSS, for which we have been waiting for appropriate submissions. We'd be very happy for you to help trial that aspect, for which we all at both rOpenSci and JSS would really appreciate if you could first wait for the review here to be completed, and then submit to JSS for (hopefully) expedited review there. Let us know if you have any questions, and @rkillick & @tdhock feel free to comment further regarding JSS alignment.
