### 2022-07-12 #549 @markean
@mpadge Thank you for the details on timeline and JSS interaction! Are there some cases (packages) that took the route from rOpenSci to JSS? It would be very helpful for us to see some good examples.
